+++ build_runner/lib/src/entrypoint/serve.dart
@@ -96,6 +96,7 @@ class ServeCommand extends WatchCommand {
       isReleaseBuild: options.isReleaseBuild,
       logPerformanceDir: options.logPerformanceDir,
       directoryWatcherFactory: defaultDirectoryWatcherFactory,
+      buildFilters: options.buildFilters,
     );
 
     if (handler == null) return ExitCode.config.code;